+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - GenerateDocumentationFile
+- Split files
 
 ## [1.0.5] - 2020-12-18
 ### Added